requirements.txt
# Image processing
Pillow==11.0.0
# Code formatting and linting
black==24.10.0
pre-commit==4.0.1
